Clean up javadoc and imit favorite suggestions to 3
